# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Ticketing support microservice",
        "title": "Gin Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Cypodsolutions",
            "url": "http://www.cypod.com/",
            "email": "naders@cypodsolutions.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/bundles": {
            "get": {
                "description": "Get bundles for a specific customer - get all system bundles if no customer specified",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundles"
                ],
                "summary": "Get Bundles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get bundle body",
                        "name": "customer_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Bundle"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new Bundle specifying its name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundles"
                ],
                "summary": "Create new Bundle",
                "parameters": [
                    {
                        "description": "Create bundle body",
                        "name": "arg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createBundleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Bundle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/bundles/assign": {
            "post": {
                "description": "Assign a bundle to a specific customer using bundle id and customer name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "bundles"
                ],
                "summary": "Assign bundle to customer",
                "parameters": [
                    {
                        "description": "Assign bundle body",
                        "name": "arg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.assignBundleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "bool"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/bundles/{bundle_id}": {
            "delete": {
                "description": "Delete  Bundle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundles"
                ],
                "summary": "Delete  Bundle",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bundle ID",
                        "name": "bundle_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "true"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
        "/orders": {
            "put": {
                "description": "Update an order specifying its end date, bundle mrc, nrc flag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Update Order with actual params",
                "parameters": [
                    {
                        "description": "Create Order body",
                        "name": "arg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateOrderRequestJSON"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "order ID for update",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create a new order specifying its batch name, it will create with no NRC, default MRC for 1 year",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create new Order",
                "parameters": [
                    {
                        "description": "Create Order body",
                        "name": "arg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.assignBundleRequest": {
            "type": "object",
            "required": [
                "bundle_id",
                "customer_name"
            ],
            "properties": {
                "bundle_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "customer_name": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "api.createBundleRequest": {
            "type": "object",
            "required": [
                "description",
                "mrc"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "mrc": {
                    "type": "number",
                    "minimum": 1
                }
            }
        },
        "api.createOrderRequest": {
            "type": "object",
            "required": [
                "batch_name"
            ],
            "properties": {
                "batch_name": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "api.updateOrderRequestJSON": {
            "type": "object",
            "required": [
                "bundle_id",
                "end_date"
            ],
            "properties": {
                "bundle_id": {
                    "type": "integer",
                    "minimum": 1
                },
                "end_date": {
                    "type": "string"
                },
                "nrc": {
                    "$ref": "#/definitions/sql.NullFloat64"
                },
                "start_date": {
                    "$ref": "#/definitions/sql.NullTime"
                }
            }
        },
...
        "db.Bundle": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mrc": {
                    "type": "number"
                }
            }
        },
...
        "db.Order": {
            "type": "object",
            "properties": {
                "batch_id": {
                    "type": "integer"
                },
                "bundle_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nrc": {
                    "$ref": "#/definitions/sql.NullFloat64"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "sql.NullFloat64": {
            "type": "object",
            "properties": {
                "float64": {
                    "type": "number"
                },
                "valid": {
                    "description": "Valid is true if Float64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}